[FE/#8] Refactor : 테이블 헤더 내용 Toolbar 파일 분리
